move `check_warnings` to a new `testing` module
