+++ src/pyuvdata/testing/__init__.py
@@ -1,0 +1,7 @@
+# -*- mode: python; coding: utf-8 -*-
+# Copyright (c) 2024 Radio Astronomy Software Group
+# Licensed under the 2-clause BSD License
+
+"""Helpful testing functions."""
+
+from .warning_check import check_warnings  # noqa
