+++ tests/test_ms_utils.py
@@ -10,8 +10,7 @@ import pytest
 
 from pyuvdata import ms_utils
 from pyuvdata.data import DATA_PATH
-
-from . import check_warnings
+from pyuvdata.testing import check_warnings
 
 pytest.importorskip("casacore")
